refactor logger and blog model
