+++ part4/bloglist/index.js
@@ -4,29 +4,27 @@ const app = express();
 const cors = require('cors');
 const mongoose = require('mongoose');
 
+const Blog = require('./models/blog');
+
 const config = require('./utils/config');
-
-const blogSchema = new mongoose.Schema({
-  title: String,
-  author: String,
-  url: String,
-  likes: Number,
-});
-
-const Blog = mongoose.model('Blog', blogSchema);
+const logger = require('./utils/logger');
 
 const mongoUrl = config.MONGODB_CONNECTION_STRING;
-if (!mongoUrl) {
-  console.error('MongoDB connection string not defined');
-  process.exit(1);
-}
 
-mongoose.connect(mongoUrl, {
-  useNewUrlParser: true,
-  useUnifiedTopology: true,
-  useFindAndModify: false,
-  useCreateIndex: true,
-});
+mongoose
+  .connect(mongoUrl, {
+    useNewUrlParser: true,
+    useUnifiedTopology: true,
+    useFindAndModify: false,
+    useCreateIndex: true,
+  })
+  .then(() => {
+    logger.info('connected to MongoDB');
+  })
+  .catch((error) => {
+    logger.error('error connecting to MongoDB:', error.message);
+    process.exit(1);
+  });
 
 app.use(cors());
 app.use(express.json());
