+++ part4/bloglist/utils/logger.js
@@ -1,0 +1,12 @@
+const info = (...params) => {
+  console.log(...params);
+};
+
+const error = (...params) => {
+  console.error(...params);
+};
+
+module.exports = {
+  info,
+  error,
+};
